fix: header, navigator, footer 폴더구조 정리, market detail 상세 정보 조회 가능하도록 수정
