+++ src/components/Layout/index.tsx
@@ -1,0 +1,3 @@
+export default function Layout() {
+    return <>헤더랑 네비게이터를 여기서 처리</>;
+}
